AtmosphericExtinction

Documentation for AtmosphericExtinction.

AtmosphericExtinction

AtmosphericExtinction.RecipeData

AtmosphericExtinction.ExtinctSpectrum

AtmosphericExtinction.ExtinctionLimits

AtmosphericExtinction.ExtinctionValues

AtmosphericExtinction.GetKnownRecipes

AtmosphericExtinction.Mag2Lin

AtmosphericExtinction.ReadData

AtmosphericExtinction.RecipeData — Type

struct RecipeData

#Arguments

table: function giving the table interpolation

lims: minimum and maximum wavelengths covered by the inout data.

AtmosphericExtinction.ExtinctSpectrum — Method

ExtinctSpectrum(recipe,iwave,ispec,ierrspec,airmass)

#Arguments

recipe is one of the extinction tables available in the package (i.e., those reported by GetKnownRecipes()).

iwave is the vector of input wavelengths. They can have units (or assumed to be Angstrom).

ispec is the vector of the input spectrum, in linear units.

ierrspec is the vector of the uncertainties associated to the input spectrum. It cna well be a vector of zeros.

airmass is the airmass associated to the observation.

Returns the input spectrum (in linear units) after atmospheric extinction.

AtmosphericExtinction.ExtinctionLimits — Method

ExtinctionLimits(recipe)

#Arguments

recipe is one of the extcintion tables available in the package (i.e., those reported by GetKnownRecipes()).

Returns the extinction values wavelength limits in Angstrom.

Examples

ExtinctionLimits("Paranal.dat")

AtmosphericExtinction.ExtinctionValues — Method

ExtinctionValues(recipe,waverange)

#Arguments

recipe is one of the extinction tables available in the package (i.e., those reported by GetKnownRecipes()).

waverange is a vector of the wavelengths of interest (in Angstrom).

Returns the extinction values for a given wavelength range in Angstrom.

Examples

ExtinctionValues("Paranal.dat",3500:3600)

AtmosphericExtinction.GetKnownRecipes — Method

GetKnownRecipes()

Returns the atmospheric extinction currently supported by the package.

Examples

GetKnownRecipes()

AtmosphericExtinction.Mag2Lin — Method

Mag2Lin(excnt)

#Arguments

extcnt is a vector with extinction values in magnitudes (per airmass, usually).

Returns the extinction values in linear units

Examples

Mag2Lin([0.5,0.4,0.3])
3-element Vector{Float64}:
 0.6309573444801932
 0.6918309709189364
 0.7585775750291838

AtmosphericExtinction.ReadData — Method

ReadData(fname)

#Arguments

fname is the file nama of a given extinction table.

Returns data for an atmospheric extinction currently supported by the package.

Examples

ReadData("Paranal.dat")